[jlox-in-cpp] Have a linked list of environments
I'd designed my environments to be a vector instead of linked lists, but
when we implement functions in Chapter 9, the linked list actually
becomes necessary. Implement that bit as a separate change to make the
actual function implementation change more focused. shared_ptr might
seem like a strange choice right now but it's what we want for
implementing functions.