[jlox-in-cpp] Add ref-counted StringValue
It's okay to keep tokens and AST nodes allocated for the lifetime of the
interpreter, since they're bounded by the input program size, but now
that we have loops, strings generated at runtime through concatenation
can grow unbounded, and keeping them around permanently is potentially
very inefficient. Add a simple ref-counted StringValue type instead so
that strings are deallocated when they become unused.

With the following test program:

    var s = "s";
    for (var i = 0; i < 100000; i = i + 1)
      s = s + "s";

With the previous implementation, it took 0.9 seconds and used 5.2 GB of
memory. It now takes 0.1 seconds and uses 1.9 MB of memory.

Adding other assignments afterwards (e.g. `var s2 = s`) doesn't increase
memory usage, confirming the ref-counting is working as intended. I also
printed out some (smaller) strings to verify their contents.

If I add one more zero:

    var s = "s";
    for (var i = 0; i < 1000000; i = i + 1)
      s = s + "s";

The interpreter takes 9.6 seconds and 16.7 MB of memory, so the time
growth isn't linear but the memory growth is. Interestingly enough, if I
convert the same program to C++:

    #include <string>
    std::string test() {
      std::string s;
      for (int i = 0; i < 1000000; i = i + 1)
        s = s + "s";
      return s;
    }

And compile it with -O3 (with some separate compilation tricks to make
sure the compiler doesn't throw away the whole loop), it takes 10.3
seconds and uses 15.7 MB of memory, so we're basically on par here. (If
you change the loop to a += then the runtime goes down to 0.2 seconds
and the memory usage to 3.7 MB, but that's less fun.)