[jlox-in-cpp] Build with C++20
I was using C++17 because of how incomplete libc++'s C++20 support is on
Xcode 13.2 (which I'm limited to for OS version reasons), but it does
implement heterogenous lookups in unordered containers, so C++20 it is.
Make a C++ standard check be a libc++ version check instead to
accommodate this.