[jlox-in-cpp] Add simple test runner
I've been wanting to do this for a while, and I discovered a bunch of
memory bugs in my REPL which finally pushed me to add regression tests.
It'd be nice to have support for tests which are expected to produce
parse or runtime errors as well; I'll add support for that later.