[jlox-in-rust] Add Expr
Sum types are super nice for this :) std::variant is passable, but Rust
is much more ergonomic here.

Unlike in C++, where each AST node just had a pointer to other nodes and
didn't own them, I'm using Box in Rust, which is an owning reference.
The T in AST does stand for tree after all, so the ownership works out.
It forces nodes to be heap-allocated, but that was always the case in
the parser anyway. I'm not sure how to express the pattern I used in C++
where an external vector owned and tracked the AST nodes; I suspect it
would involve a lot of fighting the borrow checker, which feels
pointless given that Box is perfectly serviceable here.

Having both Expr::Literal and expr::Literal (as well as token::Literal)
is maybe not the clearest thing. I'll re-evaluate that as I write the
parser and might reconsider my naming choices.