[jlox-in-rust] Add AstPrinter
This demonstrates working with the AST. In my C++ implementation, I kept
AstPrinter around till the end, but it was pretty pointless and just got
in the way. I'll just revert this right away to avoid the same fate, but
it was good practice for working with the AST (and multiple binaries in
Cargo).