[jlox-in-rust] Remove else after return. NFC
I generally like the LLVM coding standards wherever they apply, and
https://llvm.org/docs/CodingStandards.html#don-t-use-else-after-a-return
covers this case. Rust is expression-oriented, so it's a little trickier
to think about this, but the guideline I've settled down on is "use an
if/else if you'd use a ternary in C++, otherwise omit the else", so
follow that here.