[jlox-in-rust] Rename token_type to r#type
I can't use type directly because it's a Rust keyword, but I've settled
on using r# in other places where that applies (e.g. match). It's less
typing than the token_ prefix, and that prefix was also redundant in
most places since it was obvious you were referring to a token, so this
is cleaner.

VSCode's rename functionality (with rust-analyzer) Just Worked for this,
which was *so* nice.