[jlox-in-rust] Make invalid tokens unrepresentable
I was storing the token's type and literal separately, which meant that
you had the potential for invalid combinations that had to be checked at
runtime, which is silly. Make the token's type be a proper sum type with
associated literals, so that it's impossible to create an invalid token.
This is where proper sum types (vs. std::variant) really shine.

This does mean that the string representation of a token diverges from
jlox, but that was only for debugging purposes anyway. Remove the string
representation since it's unused now; I can always add #[derive(Debug)]
again if needed.

Some token_type parameters are renamed to r#type here. This should have
been part of the previous commit, but I just refactored the struct field
name and forgot about other uses. That's the downside of using proper
refactoring versus blind search and replace, I guess :D

Making functions take &TokenType instead of TokenType is premature
optimization on my part. It just feels gross to take a large type by
value, although the Rust ABI just passes via reference anyway:
https://godbolt.org/z/G3n4eesTd. I'm not sure if that's guaranteed
though, so I wanted to be explicit. The hope is that all of this gets
inlined to just the discriminator checks anyway and we never create
actual TokenType instances just for checking purposes.