[jlox-in-rust] Make invalid binary/unary expressions unrepresentable
Binary and unary expression AST nodes must be associated with an
appropriate type of token. I was previously storing a generic token for
them, but then my interpreter would have had to assert that the token
type was expected, which is unfortunate. I also wouldn't have gotten any
compile-time exhaustiveness checking for handling all operators, which
was even worse.

Introduce sub-enums for the specific token types that can appear in
binary and unary expressions, and specific token structs to encapsulate
those types alongside a line number, which is all we need for runtime
error reporting; I again benefited from having my C++ implementation at
hand to know this would be okay. The one wrinkle is needing to compute
the lexeme in the AstPrinter, but that's straightforward enough, and
I'll get rid of the AstPrinter once evaluation is in place anyway, so
it's just a temporary annoyance.