[clox] Implement strings and string operations
This completes Chapters 19.1 through 19.4.

I'm using a flexible array member instead of a separate heap-allocated
string, as suggested in Challenge 19.1. This precludes Challenge 19.2,
where we avoid heap-allocating literals, but from my experience with
jlox-in-cpp I think that optimization is more trouble than its worth
(and I'm not implementing it in jlox-in-rust either).

I'm also reworking BINARY_OP a bit to enable reusing it for the
arithmetic portion of OP_ADD, rather than needing to duplicate its code
for that.

Making objVal a function instead of a macro turns out to be a bit
annoying here, since we have to cast the argument to `Obj *` explicitly.
I'll think about how to improve that later.