[clox] Implement locals
This completes Chapter 22. Getting and setting locals was the last part
of the chapter, and you can't do anything interesting with them before
that, so the whole chapter is a single commit this time. This mostly
follows the book, except for my use of unsigned instead of signed
variables allowing a completely pointless tiny optimization.